ex1 House test class
Define the House test class. Its test fixture creates a little house with a chimney and opens a dia-
log with the House prepared message. Name the variables into which you will give individual
parts of the little house as building, chimney and roof.
Then define the test method testSmoke in this class which will supplement the smoke. Call
the individual smoke objects (ellipses) as s1, s2 and s3.