Object Type Parameters, ex Face
Define the Face class. Its test fixture draws a rectangle face and opens a dialog with the Face
prepared message. Name the variables, into which you will put individual parts as head, leftEye,
rightEye, nose and mouth.
Define the testSmile test method in this class, which evokes a smile by enlarging the height
of the mouth ellipse and covering its upper half by a rectangle. Name the variable into which
you will give the reference covering the rectangle lip.